+++ pcc_server.c
@@ -17,7 +17,7 @@ void SIGINT_handler(int signum);
 void print_statistics();
 
 // Global variables:
-uint16_t pcc_total[95];
+uint32_t pcc_total[95];
 int connfd = -1;
 int waiting_for_clients = 1;
 
@@ -40,11 +40,11 @@ int main(int argc, char *argv[]) {
     int total_sent, not_sent, cur_sent, message_len, i;
     int listenfd = -1;
     const int enable = 1;
-    uint16_t N, C;
+    uint32_t N, C;
     C = 0;
 
     char buffer[1000000]; // buffer with less than 1MB
-    uint16_t temp_pcc_total[95];
+    uint32_t temp_pcc_total[95];
     struct sigaction new_sigint_action = {
             .sa_handler = SIGINT_handler,
             .sa_flags = SA_RESTART};
